backend/requirements.txt
alembic
annotated-types
anyio
asyncpg
bcrypt
cachetools
certifi
cffi
chardet
charset-normalizer
click
cryptography
cssselect
cssutils
dnspython
email_validator
emails
fastapi
fastapi-cli
fastapi-cloud-cli
GeoAlchemy2
greenlet
h11
httpcore
httptools
httpx
idna
Jinja2
lxml
Mako
markdown-it-py
MarkupSafe
mdurl
more-itertools
mypy
mypy_extensions
numpy
packaging
passlib
pathspec
premailer
psycog
psycopg
pycparser
pydantic
pydantic-settings
pydantic_core
Pygments
PyJWT
python-dateutil
python-dotenv
python-multipart
PyYAML
requests
rich
rich-toolkit
rignore
sentry-sdk
shapely
shellingham
six
sniffio
SQLAlchemy
starlette
typer
types-passlib
typing-inspection
typing_extensions
urllib3
uvicorn
uvloop
watchfiles
websockets
